Implemented a particle effect for the knockback ability. Improved the Effect class by actually using the getEntitiesInImpactArea method... General ingame keyboard commands are now also available in the trainyard. Improved the GeometricUtilities. Improved particle implemntations.
git-svn-id: https://dev.wibe.media/svn/code/java/trunk/litiengine@1409 16493bb7-3577-b145-b9c3-72a97d279fb6

+++ src/de/gurkenlabs/litiengine/graphics/particles/LeftLineParticle.java
@@ -9,7 +9,7 @@ import de.gurkenlabs.litiengine.Game;
 
 public class LeftLineParticle extends Particle {
 
-  public LeftLineParticle(final float xCurrent, final float yCurrent, final float dx, final float dy, final float gravityX, final float gravityY, final byte width, final byte height, final int life, final Color color) {
+  public LeftLineParticle(final float xCurrent, final float yCurrent, final float dx, final float dy, final float gravityX, final float gravityY, final float width, final float height, final int life, final Color color) {
     super(xCurrent, yCurrent, dx, dy, gravityX, gravityY, width, height, life, color);
   }
