+++ src/de/gurkenlabs/litiengine/graphics/particles/RightLineParticle.java
@@ -9,7 +9,7 @@ import de.gurkenlabs.litiengine.Game;
 
 public class RightLineParticle extends Particle {
 
-  public RightLineParticle(final float xCurrent, final float yCurrent, final float dx, final float dy, final float gravityX, final float gravityY, final byte width, final byte height, final int life, final Color color) {
+  public RightLineParticle(final float xCurrent, final float yCurrent, final float dx, final float dy, final float gravityX, final float gravityY, final float width, final float height, final int life, final Color color) {
     super(xCurrent, yCurrent, dx, dy, gravityX, gravityY, width, height, life, color);
   }
