+++ src/de/gurkenlabs/litiengine/graphics/particles/ShimmerParticle.java
@@ -45,7 +45,7 @@ public class ShimmerParticle extends Particle {
    * @param particleType
    *          the particle type
    */
-  public ShimmerParticle(final Rectangle2D boundingBox, final float xCurrent, final float yCurrent, final float dx, final float dy, final float gravityX, final float gravityY, final byte width, final byte height, final int life, final Color color) {
+  public ShimmerParticle(final Rectangle2D boundingBox, final float xCurrent, final float yCurrent, final float dx, final float dy, final float gravityX, final float gravityY, final float width, final float height, final int life, final Color color) {
     super(xCurrent, yCurrent, dx, dy, gravityX, gravityY, width, height, life, color);
     this.boundingBox = boundingBox;
   }
